refactor(spica): rename trace adapter module
