+++ src/aiconfigurator/spica/cli_adapter.py
